커밋하지 않았던 최종 수정 코드

+++ client/src/components/myaccount/Friends.vue
@@ -6,22 +6,21 @@
           <div class="d-flex justify-content-between align-items-center">
             <h1 class="h3 m-4">팔로워 (Follower)</h1>
             <i @click="showFriendInsertModal" class="bi bi-person-plus-fill float-end cursor-pointer me-2"></i>
-            <!-- <button @click="showFriendInsertModal" style="font-size: 14px;" class="btn btn-primary mt-4 mb-4 px-2 py-1 lift float-end">친구 추가</button> -->
           </div>
           <div class="row row-cols-xl-6 row-cols-lg-3 row-cols-md-3 row-cols-sm-2 row-cols-2 g-1 mb-4 row-deck">
             <div v-for="(item, index) in followingList" :key="index" class="col">
               <div class="card text-center">
-                  <!-- Group Hover action -->
                 <div class="btn-group position-absolute top-0 right-0">
                   <a href="#" class="nav-link py-2 px-3 text-muted" data-bs-toggle="dropdown" aria-expanded="false"><i class="fa fa-ellipsis-v"></i></a>
                   <ul class="dropdown-menu dropdown-menu-end border-0 shadow">
                     <li><a @click="showFriendBookmarkModal(item.userId, item.userName)" class="dropdown-item">관심 매물 보기</a></li>
-                    <li><a @click="deleteFriend(item.userId)" class="dropdown-item">삭제</a></li>
+                    <li><a @click="deleteFriend(item.userId, item.userName)" class="dropdown-item">삭제</a></li>
                   </ul>
                 </div>
                 <div class="card-body d-flex align-items-center justify-content-between flex-column">
                   <div class="me-auto ms-auto py-4">
-                    <img src="http://localhost:8080/images/profile_av.png" alt="" class="rounded-circle">
+                    <img v-if="item.userProfileImage" :src="item.userProfileImage" :alt="item.userName" style="width: 140px; height: 140px;" class="rounded-circle">
+                    <img v-else src="http://localhost:8080/images/profile_av.png" :alt="item.userName" class="rounded-circle">
                   </div>
                   <div class="mt-2">
                       <h6 class="mb-0">{{ item.userName }}</h6>
@@ -39,18 +38,17 @@
           <div class="row row-cols-xl-6 row-cols-lg-3 row-cols-md-3 row-cols-sm-2 row-cols-2 g-1 mb-4 row-deck">
             <div v-for="(item, index) in followerList" :key="index+50" class="col">
               <div class="card text-center">
-                  <!-- Group Hover action -->
-                <div class="btn-group position-absolute top-0 right-0">
+                <!--<div class="btn-group position-absolute top-0 right-0">
                   <a href="#" class="nav-link py-2 px-3 text-muted" data-bs-toggle="dropdown" aria-expanded="false"><i class="fa fa-ellipsis-v"></i></a>
                   <ul class="dropdown-menu dropdown-menu-end border-0 shadow">
                     <li><a class="dropdown-item" href="#">관심 매물 보기</a></li>
                     <li><a class="dropdown-item" href="#">삭제</a></li>
                   </ul>
-                </div>                                            
+                </div>-->
                 <div class="card-body d-flex align-items-center justify-content-between flex-column">
                   <div class="me-auto ms-auto py-4">
-                    <img v-if="profileImgUrl" :src="profileImgUrl" alt="이미지" style="width: 140px; height: 140px;" class="rounded-circle">
-                    <img v-else src="http://localhost:8080/images/profile_av.png" alt="기본 이미지" class="rounded-circle">
+                    <img v-if="item.userProfileImage" :src="item.userProfileImage" :alt="item.userName" style="width: 140px; height: 140px;" class="rounded-circle">
+                    <img v-else src="http://localhost:8080/images/profile_av.png" :alt="item.userName" class="rounded-circle">
                   </div>
                   <div class="mt-2">
                       <h6 class="mb-0">{{ item.userName }}</h6>
@@ -103,7 +101,6 @@ export default {
     getFollowingList() {
       http.get('/friend/following')
         .then(({ data }) => {
-          console.log(data);
           this.followingList = data.list;
         })
         .catch(error => this.$swal('서버에 문제가 발생하였습니다.', { icon: 'error' }))
@@ -111,19 +108,39 @@ export default {
     getFollowerList() {
       http.get('/friend/follower')
         .then(({ data }) => {
-          console.log(data);
           this.followerList = data.list;
         })
         .catch(error => this.$swal('서버에 문제가 발생하였습니다.', { icon: 'error' }))
 
     },
-    deleteFriend(toId) {
-      http.delete(`/friend/${toId}`)
-        .then(({ data }) => {
-          console.log(data);
-          this.getFollowingList();
-        })
-        .catch(error => this.$swal('서버에 문제가 발생하였습니다.', { icon: 'error' }))
+    deleteFriend(toId, userName) {
+
+      this.$swal({
+        title: `친구 ${userName}을(를) 삭제하시겠습니까?`,
+        icon: 'warning',
+        dangerMode: true,
+        buttons: true
+      }).then(value => {
+        if (value) {
+          http.delete(`/friend/${toId}`)
+            .then(({ data }) => {
+              if (data.result == 'login') {
+                this.$swal('세션이 만료되었거나, 로그인되지 않았습니다. 로그인 페이지로 이동합니다.', { icon: 'warning' })
+                  .then(() => {
+                    this.SET_USER_LOGOUT();
+                    this.$router.push('/user/login');
+                  })
+              } else {
+                this.$swal('삭제 완료되었습니다.', { icon: 'success' })
+                  .then((value) => {
+                    this.getFollowingList();
+                  }
+                );
+              }
+            })
+            .catch(error => this.$swal('서버에 문제가 발생하였습니다.', { icon: 'error' }))
+        }
+      })
     },
     showFriendInsertModal() {
       this.friendInsertModal.show();
@@ -138,11 +155,11 @@ export default {
     closeAfterInsert() {
       this.friendInsertModal.hide();
       this.getFollowingList();
-      console.log('모달 종료 후 동작!');
+      // console.log('모달 종료 후 동작!');
     },
     closeAfterFriendBookmark() {
       this.friendBookmarkModal.hide();
-      console.log('북마크모달종료')
+      // console.log('북마크모달종료')
     }
   },
   mounted() {
